+++ local_datasource/src/main/java/com/baghdad/local_datasource/LocalActorDataSourceImpl.kt
@@ -5,7 +5,7 @@ import com.baghdad.local_datasource.roomDB.entity.Actor
 import com.baghdad.local_datasource.roomDB.entity.toDto
 import com.baghdad.local_datasource.roomDB.errorHandler.executeWithErrorHandling
 import com.baghdad.repository.datasource.local.LocalActorDataSource
-import com.baghdad.repository.model.ActorDto
+import com.baghdad.repository.model.actor.ActorDto
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
