+++ remote_datasource/src/main/java/com/baghdad/remoteDataSource/mapper/GenreResponseMapper.kt
@@ -12,3 +12,4 @@ fun GenreListResponse.toDto(genreType: GenreDto.GenreType): List<GenreDto> {
         )
     }
 }
+
