+++ repository/src/main/java/com/baghdad/repository/model/CastMemberDto.kt
@@ -1,5 +1,7 @@
 package com.baghdad.repository.model
 
+import com.baghdad.repository.model.actor.ActorDto
+
 data class CastMemberDto(
     val actor: ActorDto,
     val characterName: String
